fix(merge): restore int type for MergeConfig.CollectionsPerDecommit
PR #11937 ("Unify types with Geth") changed CollectionsPerDecommit
from int to ulong. This is an internal GC-tuning knob with no Geth
equivalent, and -1 is a documented, supported sentinel meaning
"never decommit" (see the ConfigItem description and the original
NoGCStrategy default of -1).

After the change, supplying -1 (used by the EXPB reproducible
benchmark and any deployment that disables decommit) fails config
binding with OverflowException, crashing the node on startup before
JSON-RPC comes up. This broke the run-expb-reproducible-benchmarks
workflow.

Revert just this field (and the IGCStrategy / NoGCStrategy /
NoSyncGcRegionStrategy / GCKeeper consumers) back to int, restoring
the -1 sentinel. Genuinely Geth-comparable fields from #11937 are
left unchanged.

+++ src/Nethermind/Nethermind.Merge.Plugin/GC/NoSyncGcRegionStrategy.cs
@@ -23,7 +23,7 @@ public class NoSyncGcRegionStrategy : IGCStrategy
         PostBlockDelayMs = mergeConfig.PostBlockGcDelayMs ?? (int)((mergeConfig.SecondsPerSlot * 1000) / 8);
     }
 
-    public ulong CollectionsPerDecommit { get; }
+    public int CollectionsPerDecommit { get; }
     public int PostBlockDelayMs { get; }
 
     public bool CanStartNoGCRegion() => _canStartNoGCRegion && _syncModeSelector.Current == SyncMode.WaitingForBlock;
